requirements.txt
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
pydantic==2.10.5
pydantic_core==2.27.2
python-slugify==8.0.4
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
text-unidecode==1.3
typing_extensions==4.12.2
uvicorn==0.34.0
